backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
passlib[bcrypt]
python-dotenv